STEP1: 폴더링, 코드 분리
