--- ParcelInvoiceMaker/ParcelInvoiceMaker/ParcelProcessor.swift
@@ -1,134 +1,0 @@
-//
-//  ParcelInvoiceMaker - ParcelProcessor.swift
-//  Created by yagom.
-//  Copyright © yagom. All rights reserved.
-// 
-
-import Foundation
-
-struct MobileNumber {
-    private let value: String
-    
-    init(value: String) {
-        self.value = value
-    }
-    
-    func getValue() -> String {
-        return self.value
-    }
-}
-
-struct Name {
-    private let value: String
-    
-    init(value: String) {
-        self.value = value
-    }
-    
-    func getValue() -> String {
-        return self.value
-    }
-}
-
-struct ReceiverInformation {
-    private let address: String
-    private let name: Name
-    private let mobileNumber: MobileNumber
-    
-    func getAddress() -> String {
-        return self.address
-    }
-    
-    func getReceiverName() -> String {
-        return self.name.getValue()
-    }
-    
-    func getReceiverMobile() -> String {
-        return self.mobileNumber.getValue()
-    }
-    
-    init(address: String, 
-         name: Name,
-         mobileNumber: MobileNumber)
-    {
-        self.address = address
-        self.name = name
-        self.mobileNumber = mobileNumber
-    }
-}
-
-final class ParcelInformation {
-    private let receiverInformation: ReceiverInformation
-    private let deliveryCost: Int
-    private let discount: Discount
-    
-    init(receiverInformation: ReceiverInformation,
-         deliveryCost: Int,
-         discount: Discount)
-    {
-        self.receiverInformation = receiverInformation
-        self.deliveryCost = deliveryCost
-        self.discount = discount
-    }
-    
-    func getReceiverName() -> String {
-        return receiverInformation.getReceiverName()
-    }
-    
-    func getReceiverAddress() -> String {
-        return receiverInformation.getAddress()
-    }
-    
-    func getReceiverMobile() -> String {
-        return receiverInformation.getReceiverMobile()
-    }
-    
-    func getDiscountedCost() -> Int {
-        switch discount {
-        case .none:
-            return deliveryCost
-        case .vip:
-            return deliveryCost / 5 * 4
-        case .coupon:
-            return deliveryCost / 2
-        }
-    }
-}
-
-enum Discount: Int {
-    case none = 0, vip, coupon
-}
-
-protocol ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation)
-}
-
-final class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation) {
-        // 데이터베이스에 주문 정보 저장
-        print("발송 정보를 데이터베이스에 저장했습니다.")
-    }
-}
-
-protocol ParcelOrderProcessable {
-    func process(parcelInformation: ParcelInformation,
-                 onComplete: (ParcelInformation) -> Void)
-}
-
-final class ParcelOrderProcessor: ParcelOrderProcessable {
-    private let persistence: ParcelInformationPersistence
-    
-    init(persistence: ParcelInformationPersistence) {
-        self.persistence = persistence
-    }
-    
-    // 택배 주문 처리 로직
-    func process(parcelInformation: ParcelInformation, 
-                 onComplete: (ParcelInformation) -> Void) {
-        
-        // 데이터베이스에 주문 저장
-        self.persistence.save(parcelInformation: parcelInformation)
-        
-        onComplete(parcelInformation)
-    }
-}
